# frontend/ticketing_app_ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/tickets": {
      "get": {
        "tags": [
          "Ticket Management"
        ],
        "summary": "Recupera tutti i ticket (paginati e filtrati per ruolo)",
        "description": "Recupera una lista paginata di ticket in base al ruolo dell'utente autenticato, con opzioni di filtro.",
        "operationId": "getTickets",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtra per stato del ticket",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ANSWERED",
                "SOLVED",
                "DRAFT"
              ]
            },
            "example": "OPEN"
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filtra per priorità del ticket",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            },
            "example": "HIGH"
          },
          {
            "name": "search",
            "in": "query",
            "description": "Cerca per titolo, descrizione, nome categoria o nome servizio",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "problema"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista di ticket recuperata con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTicketResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTicketResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (ruolo non supportato).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTicketResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ticket Management"
        ],
        "summary": "Crea un nuovo ticket o aggiorna una bozza esistente",
        "description": "Permette a USER, HELPER, PM e ADMIN di creare un nuovo ticket o di salvare/finalizzare una bozza.",
        "operationId": "createOrUpdateTicket",
        "parameters": [
          {
            "name": "ticketId",
            "in": "query",
            "description": "ID del ticket da aggiornare (solo per le bozze)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "201": {
            "description": "Ticket creato/aggiornato con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dati di input non validi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Risorsa non trovata (es. categoria, servizio).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tickets/{ticketId}/status": {
      "patch": {
        "tags": [
          "Ticket Management"
        ],
        "summary": "Aggiorna lo stato di un ticket",
        "description": "Permette a HELPER, PM e ADMIN di cambiare lo stato di un ticket. Le bozze non possono essere modificate qui.",
        "operationId": "updateTicketStatus",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID del ticket da aggiornare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "newStatus",
            "in": "query",
            "description": "Nuovo stato del ticket",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ANSWERED",
                "SOLVED",
                "DRAFT"
              ]
            },
            "example": "SOLVED"
          }
        ],
        "responses": {
          "404": {
            "description": "Ticket non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Stato del ticket aggiornato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Stato non valido o tentativo di modificare una bozza.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tickets/{ticketId}/reject": {
      "patch": {
        "tags": [
          "Ticket Management"
        ],
        "summary": "Rifiuta e riassegna un ticket",
        "description": "Un helper/PM/admin rifiuta un ticket e lo riassegna a un altro helper/admin. Il ticket deve essere in stato OPEN.",
        "operationId": "rejectTicket",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID del ticket da rifiutare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "newAssignedToId",
            "in": "query",
            "description": "ID del nuovo utente (helper/admin) a cui assegnare il ticket",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "uuid-helper-2"
          }
        ],
        "responses": {
          "400": {
            "description": "Stato del ticket o nuovo assegnatario non validi.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato (non sei l'assegnatario).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Ticket o nuovo assegnatario non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Ticket rifiutato e riassegnato con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tickets/{ticketId}/assign": {
      "patch": {
        "tags": [
          "Ticket Management"
        ],
        "summary": "Assegna un ticket a un helper/admin",
        "description": "Permette a PM e ADMIN di assegnare manualmente un ticket a un utente specifico (solo helper/admin).",
        "operationId": "assignTicket",
        "parameters": [
          {
            "name": "ticketId",
            "in": "path",
            "description": "ID del ticket da assegnare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "helperId",
            "in": "query",
            "description": "ID dell'helper/admin a cui assegnare il ticket",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "uuid-helper-3"
          }
        ],
        "responses": {
          "404": {
            "description": "Ticket o helper non trovato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "ID helper non valido o tentativo di assegnare una bozza.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "Non autorizzato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Ticket assegnato con successo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageTicketResponseDTO": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketResponseDTO"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortObject"
            }
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string"
          },
          "nullHandling": {
            "type": "string"
          },
          "ascending": {
            "type": "boolean"
          },
          "property": {
            "type": "string"
          },
          "ignoreCase": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "keycloak": {
        "type": "oauth2",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "password": {
            "authorizationUrl": "http://localhost:9090/realms/ticket_tool/protocol/openid-connect/auth",
            "tokenUrl": "http://localhost:9090/realms/ticket_tool/protocol/openid-connect/token",
            "scopes": {

            }
          }
        }
      }
    }
  }
}